requirements.txt
Django==5.2.8
itemadapter==0.12.2
python-decouple==3.8
scrapy==2.13.3
psycopg2-binary==2.9.11
djangorestframework==3.15.1